adding aliases and refactoring file related to init command
