+++ src/app.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 
 import { Command } from 'commander';
-import { initCommand } from './commands/initCommand';
+import { initCommand } from './commands/init';
 
 const program = new Command();
