+++ src/helpers/promptsHelper.ts
@@ -1,6 +1,7 @@
 import prompts from 'prompts';
-import { UserAnswers } from '../interface/UserConfiguration';
-import { exitCommand as onCancel } from '../commands/exitCommand';
+
+import { UserAnswers } from '@/interfaces/UserConfiguration';
+import { exitCommand as onCancel } from '@/commands/exitCommand';
 
 export function configPrompt<T extends string>(questions: prompts.PromptObject<T> | Array<prompts.PromptObject<T>>, options?: prompts.Options): Promise<UserAnswers> {
   return prompts<T>(questions, { onCancel }) as Promise<UserAnswers>;
