+++ src/interfaces/UserConfiguration.ts
