+++ src/models/index.ts
@@ -1,3 +1,4 @@
 export * from './ConfirmQuestion';
 export * from './NumberQuestion';
 export * from './ToggleQuestion';
+export * from './SelectQuestion'
